+++ eco-core/core-plugin/src/main/java/com/willfp/ecoskills/api/modifier/StatModifier.java
@@ -23,7 +23,7 @@ public interface StatModifier {
      *
      * @return The amount.
      */
-    int getAmount();
+    double getAmount();
 
     /**
      * Get the operation.
